+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Unreleased
+## [1.8.0] - 2024-06-24
 ### Added
 * Added document minification as a feature before document translation, to allow
 translation of large `docx` or `pptx` files. For more info check the README.
@@ -154,7 +154,9 @@ Stable release.
 Initial version.
 
 
-[1.7.2]: https://github.com/DeepLcom/deepl-php/compare/v1.7.1...v1.7.2 
+[Unreleased]: https://github.com/DeepLcom/deepl-php/compare/v1.8.0...HEAD
+[1.8.0]: https://github.com/DeepLcom/deepl-php/compare/v1.7.2...v1.8.0
+[1.7.2]: https://github.com/DeepLcom/deepl-php/compare/v1.7.1...v1.7.2
 [1.7.1]: https://github.com/DeepLcom/deepl-php/compare/v1.7.0...v1.7.1
 [1.7.0]: https://github.com/DeepLcom/deepl-php/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/DeepLcom/deepl-php/compare/v1.5.1...v1.6.0